Squashed commit of the following:
commit e2968d3cfc2be2004371fc311654719c9423b0f2
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Thu Apr 15 08:53:59 2021 +0100

    change langue index

commit dbb70742fd382a9f86a1bd1b6088229e1fde8392
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Wed Apr 14 14:39:16 2021 +0100

    add logout

commit 037c417e5090b33cec09b67dda5373a762e5c8d4
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Wed Apr 14 12:31:14 2021 +0100

    add filters

commit f9483c3118396925b30aeba25a92ac6450630196
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Tue Apr 13 13:25:27 2021 +0100

    fix error

commit ec73de2a77f22a7fc6e1f9b80655c1df86718f2f
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Tue Apr 13 12:13:04 2021 +0100

    add experinces

commit e0026526f55bcbf54eae38c3f062a0855b5f9577
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Wed Apr 7 12:22:21 2021 +0100

    add loading download pdf

commit 23998304a4254e62feb43c3d75442e85e263c7d7
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Tue Apr 6 17:52:21 2021 +0100

    add groupe list

commit c3e108acd4848f5b5eff13dfb1c0f639d28bf45b
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Tue Apr 6 17:19:59 2021 +0100

    add refCall

commit 55c88e21d526f12345478e15295cf460ef1b28d9
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Tue Apr 6 17:03:20 2021 +0100

    add crud exps

commit 07d7470ad4f513edccabafde2a8987df61fea22c
Author: mahdihellal <mahdihellal@gmail.com>
Date:   Tue Apr 6 12:39:15 2021 +0100

    add exp container

+++ src/components/Layout/Drawer/Drawer.tsx
@@ -3,6 +3,7 @@
 import React, { useState, useContext, useEffect } from 'react';
 import DrawerContext from 'common/contexts/DrawerContext';
 import { setAuthorizationBearer, client } from 'common/requests/client';
+import { useLogout } from 'common/requests/auth';
 import userContext from 'common/contexts/UserContext';
 import localforage from 'localforage';
 import drawerLogo from 'assets/svg/drawerLogo.svg';
@@ -25,6 +26,7 @@ const PrivateDrawer = () => {
   const location = useLocation();
   const initialSelected = () => Number(window.localStorage.getItem('selectedButton') || -1);
   const [selectedButton, setSelectedButton] = useState(initialSelected);
+  const [logoutCall, logoutState] = useLogout();
 
   const { open, setOpen } = useContext(DrawerContext);
   const { setUser, user } = useContext(userContext);
@@ -40,12 +42,18 @@ const PrivateDrawer = () => {
     window.localStorage.setItem('selectedButton', JSON.stringify(selectedButton));
   }, [selectedButton]);
   const logout = () => {
-    localforage.removeItem('auth');
-    setAuthorizationBearer('');
-    setUser(null);
-    localStorage.clear();
-    client.clearStore();
+    logoutCall();
   };
+  useEffect(() => {
+    if (logoutState.data) {
+      localforage.removeItem('auth');
+      setAuthorizationBearer('');
+      setUser(null);
+      localStorage.clear();
+      client.clearStore();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [logoutState.data]);
   let Links = [
     {
       text: 'Parcours',
